Added event handling and finished alias and custom domain functionality.

+++ src/main/kotlin/com/criptext/mail/utils/generaldatasource/workers/UpdateLocalDomainAndAliasDataWorker.kt
@@ -8,6 +8,8 @@ import com.criptext.mail.db.dao.AliasDao
 import com.criptext.mail.db.dao.CustomDomainDao
 import com.criptext.mail.db.models.ActiveAccount
 import com.criptext.mail.db.models.Alias
+import com.criptext.mail.db.models.Contact
+import com.criptext.mail.db.models.CustomDomain
 import com.criptext.mail.scenes.settings.aliases.data.AliasData
 import com.criptext.mail.utils.UIMessage
 import com.criptext.mail.utils.generaldatasource.data.GeneralResult
@@ -17,12 +19,12 @@ import com.github.kittinunf.result.Result
 class UpdateLocalDomainAndAliasDataWorker(private val customDomainDao: CustomDomainDao,
                                           private val aliasDao: AliasDao,
                                           private val activeAccount: ActiveAccount,
-                                          private val customDomains: List<String>,
+                                          private val customDomains: List<CustomDomain>,
                                           private val aliasData: List<AliasData>,
                                           override val publishFn: (GeneralResult.UpdateLocalDomainAndAliasData) -> Unit
                           ) : BackgroundWorker<GeneralResult.UpdateLocalDomainAndAliasData> {
 
-    override val canBeParallelized = true
+    override val canBeParallelized = false
 
     override fun catchException(ex: Exception): GeneralResult.UpdateLocalDomainAndAliasData {
         return when(ex){
@@ -39,13 +41,27 @@ class UpdateLocalDomainAndAliasDataWorker(private val customDomainDao: CustomDom
 
             val newAliases = aliasData.filter { it.rowId !in dbAliases.map { dbAlias -> dbAlias.rowId } }
             val deletedAliases = dbAliases.filter { it.rowId !in aliasData.map { serverAlias -> serverAlias.rowId } }
-            aliasDao.deleteAll(deletedAliases)
-            aliasDao.insertAll(newAliases.map { Alias(0, rowId = it.rowId,
-                    accountId = activeAccount.id, domain = it.domain, active = it.isActive, name = it.name) })
+            if(deletedAliases.isNotEmpty()) {
+                aliasDao.deleteAll(deletedAliases)
+            }
+            if(newAliases.isNotEmpty()) {
+                aliasDao.insertAll(newAliases.map {
+                    Alias(0, rowId = it.rowId,
+                            accountId = activeAccount.id, domain = if (it.domain == Contact.mainDomain) null else it.domain, active = it.isActive, name = it.name)
+                })
+            }
 
-            val newDomains = customDomains.filter { it !in dbCustomDomains.map { dbDomain -> dbDomain.name } }
-            val deletedDomains = dbCustomDomains.filter { it.rowId !in aliasData.map { serverAlias -> serverAlias.rowId } }
-            customDomainDao.deleteAll(deletedDomains)
+            val newDomains = customDomains.filter { it.name !in dbCustomDomains.map { dbDomain -> dbDomain.name } }
+            val deletedDomains = dbCustomDomains.filter { (it.name !in customDomains.map { serverCustomDomain -> serverCustomDomain.name }) && (it.validated) }
+            if(deletedDomains.isNotEmpty()) {
+                customDomainDao.deleteAll(deletedDomains)
+            }
+            if(newDomains.isNotEmpty()) {
+                customDomainDao.insertAll(newDomains.map {
+                    CustomDomain(0, rowId = it.rowId,
+                            accountId = activeAccount.id, validated = it.validated, name = it.name)
+                })
+            }
         }
         return when (operation){
             is Result.Success -> {
